Update file management

+++ cmdbox/scaffold_templates/views.py
@@ -169,15 +169,18 @@ def duplicate_file(request, username, slug, file_id):
 
 
 @login_required
-@require_POST
 def delete_file(request, username, slug, file_id):
     try:
         _file = File.objects.get(pk=file_id, template__user__username=username, template__slug=slug)
-        _file.delete()
         json_context = dict()
-        files = _file.template.files.all()
-        json_context['html'] = walk(files)
-        json_context['itemsCount'] = files.count()
+        if request.method == 'POST':
+            _file.delete()
+            files = _file.template.files.all()
+            json_context['file'] = file_id
+            json_context['itemsCount'] = files.count()
+        else:
+            context = Context({'file': _file})
+            json_context['html'] = render_to_string('scaffold_templates/partial_delete_file.html', context)
         return HttpResponse(json.dumps(json_context), content_type='application/json')
     except File.DoesNotExist:
         return HttpResponseBadRequest()
